// 04-country-app/src/app/countries/interfaces/cache-store.interface.ts
import { Country } from './country.interface';
import { ValidRegion } from './region.type';

export interface CacheStore {
  byCapital: TermCountries;
  byCountry: TermCountries;
  byRegion: RegionCountries;
}

export interface TermCountries {
  term: string;
  countries: Country[];
}

export interface RegionCountries {
  region?: ValidRegion;
  countries: Country[];
}